Refactor web client snapshot tests
