+++ packages/frontend/web-client/src/components/Avatar/index.test.ts
@@ -1,0 +1,13 @@
+import { describeWithProps } from "test-utilities/react";
+
+import Avatar from ".";
+
+function describeSize (size: string) {
+  describeWithProps(size, Avatar, { size });
+}
+
+describe("Avatar component", () => {
+  describeSize("small");
+
+  describeSize("large");
+});
